move generate geolocation into lib folder
