+++ lib/artwork-scraper.js
@@ -1,7 +1,6 @@
 const readline = require('readline');
 const fs = require('fs');
-const setGeolocationForArtworks = require('../js/generate-geolocation');
-const CONSTANTS = require('../js/constants');
+const setGeolocationForArtworks = require('../lib/generate-geolocation');
 
 class ArtworkScraper {
   constructor(pathOutputFile, host, url, httpBody) {
