+++ lib/generate-geolocation.js
